Split vgpu check
